requirements2.txt
asgiref==3.3.4
Django==3.2.5
pillow
django-widget-tweaks
django-bootstrap-modal-forms
WeasyPrint
django-chartjs
gunicorn
psycopg2-binary
pep8
pylint
requests
six
mccabe
typed-ast
astroid
autopep8
certifi
chardet
django-crispy-forms
django-tinymce4-lite
idna
isort
jsmin
lazy-object-proxy
numpy
twilio
setuptools
boto3
django-storages
python-decouple
django-session-timeout
certbot
#python3-certbot-nginx
beautifulsoup4
django-crontab
django-cron


